intel-worker/requirements.txt
websockets
pyzmq
google-generativeai
MetaTrader5
feedparser
requests
pydantic
